reorg to remove CCPhyloSolver class
