+++ cassiopeia/solver/__init__.py
@@ -12,7 +12,6 @@ from .SpectralSolver import SpectralSolver
 from .UPGMASolver import UPGMASolver
 from .VanillaGreedySolver import VanillaGreedySolver
 from .SpectralNeighborJoiningSolver import SpectralNeighborJoiningSolver
-from .CCPhyloSolver import CCPhyloSolver
 from .DynamicNeighborJoiningSolver import DynamicNeighborJoiningSolver
 from .HeuristicNeighborJoiningSolver import HeuristicNeighborJoiningSolver
 from . import dissimilarity_functions as dissimilarity
